chore: Move SqlColumnHelper
